fix: remove redundant type conversions in PdfiumCore.kt

+++ src/main/java/com/hyntix/pdfium/PdfiumCore.kt
@@ -146,11 +146,11 @@ class PdfiumCore {
     }
 
     fun getPageWidth(pagePtr: Long): Double {
-        return nativeGetPageWidth(pagePtr).toDouble()
+        return nativeGetPageWidth(pagePtr)
     }
 
     fun getPageHeight(pagePtr: Long): Double {
-        return nativeGetPageHeight(pagePtr).toDouble()
+        return nativeGetPageHeight(pagePtr)
     }
 
     /**
